.claude/skills/tavily-search/requirements.txt
tavily-python>=0.7.0
requests>=2.28.0